Update generateTrainedData.bat
commend out shapeclustering based on the https://github.com/tesseract-ocr/tesseract/wiki/Training-Tesseract-3.03%E2%80%933.05#training-procedure 
==========
shapeclustering should not normally be used except for the Indic languages. (It may be necessary to use shapeclustering for Latin-alphabet languages where base characters are combined with two or more non-spacing modifiers, e.g. ḗ, ɔ̄́, r̥̄: this includes, but may not be limited to, Indic languages in Latin transcription. Failure to do so will lead mftraining to give an Assert failed:in file unicharset.cpp error.)
==========